requirements.txt
streamlit
dotenv
pymongo
google-generativeai
bcrypt
captcha